pyproject.toml
[project]
name = "pdm-bench"
version = "0.1.3"
requires-python = ">=3.10,<3.13"
description = "Benchmark tooling for cross-domain bearing fault diagnosis on CWRU and PU."
authors = [
  { name = "Pawel Knap", email = "pknap@agh.edu.pl" },
  { name = "Urszula Jachymczyk", email = "ujachymczyk@agh.edu.pl" }
]
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.0",
    "numpy>=2.0.0",
    "pandas>=2.1.2",
    "scikit-learn>=1.6.1",
    "scikit-optimize>=0.10.2",
    "scipy>=1.11.4",
    "torch>=2.2.0",
    "xgboost>=3.0.0",
]

[project.optional-dependencies]
tracking = [
    "mlflow>=3.10.0",
]

[dependency-groups]
dev = [
    "pre-commit>=3.6.0",
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.12.7",
]

[tool.ruff]
extend-exclude = [
    "docs/*",
    "dist/*",
    "notebooks/*",
    "sandbox/*",
    "scripts/*",
]
fix = true
line-length = 88
src = ["src"]
target-version = "py310"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff.lint]
fixable = ["I"]
extend-select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ERA", # flake8-eradicate/eradicate
    "I",   # isort
    "N",   # pep8-naming
    "PIE", # flake8-pie
    "PGH", # pygrep
    "RUF", # ruff checks
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "UP",  # pyupgrade
    #"D",   # pydocstyle - google
]
ignore = ["N806", "N803", "RUF100", "C420"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.coverage.run]
branch = true
# Define the source to measure—adjust this to your main package/module
source = ["src"]

[tool.coverage.report]
omit = [
    "*/__init__.py",  # Omit the main entry file, if present
    "*/tests/*"        # Omit any files in the tests directory
]

exclude_lines = [
    "if __name__ == .__main__.:",
    "def demo\\("
]

[tool.pytest.ini_options]
norecursedirs = [".*", "__pycache__", "venv", "env", "sandbox"]
markers = [
    "smoke: lightweight end-to-end smoke tests",
]
